refactor: 重新设计错误码
