+++ laokou-common/laokou-common-i18n/src/main/resources/i18n/message_zh_CN.properties
@@ -7,3 +7,12 @@
 503=服务正在维护，请联系管理员
 1001=请求过于频繁，请稍后再试
 1002=服务未知错误
+2001=唯一标识不能为空
+2002=验证码不能为空
+2003=验证码错误，请重新输入
+2004=验证码已过期
+2005=账号不能为空
+2006=账号已被停用
+2007=帐户或密码错误，请重新输入
+2008=用户没有权限访问，请联系管理员
+2009=密码不能为空
