+++ laokou-service/laokou-admin/laokou-admin-server/pom.xml
@@ -16,10 +16,6 @@
 			<artifactId>spring-boot-admin-starter-client</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>org.apache.shardingsphere</groupId>
-			<artifactId>shardingsphere-jdbc-core</artifactId>
-		</dependency>
-		<dependency>
 			<groupId>org.laokou</groupId>
 			<artifactId>laokou-oss-client</artifactId>
 		</dependency>
