+++ laokou-service/laokou-auth/laokou-auth-client/pom.xml
@@ -16,6 +16,10 @@
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-oauth2-jose</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.laokou</groupId>
+            <artifactId>laokou-common-i18n</artifactId>
+        </dependency>
         <!--安全模块-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
