+++ laokou-service/laokou-auth/laokou-auth-client/src/main/java/org/laokou/auth/client/exception/InvalidAuthenticationEntryPoint.java
@@ -17,8 +17,8 @@ package org.laokou.auth.client.exception;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.laokou.common.core.utils.MessageUtil;
-import org.laokou.common.core.exception.ErrorCode;
+import org.laokou.common.i18n.core.StatusCode;
+import org.laokou.common.i18n.utils.MessageUtil;
 import org.springframework.security.authentication.InsufficientAuthenticationException;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
@@ -44,7 +44,7 @@ public class InvalidAuthenticationEntryPoint implements AuthenticationEntryPoint
             CustomAuthExceptionHandler.handleException(response, Integer.valueOf(errorCode), message);
         }
         if (authException instanceof InsufficientAuthenticationException) {
-            CustomAuthExceptionHandler.handleException(response, ErrorCode.UNAUTHORIZED, MessageUtil.getMessage(ErrorCode.UNAUTHORIZED));
+            CustomAuthExceptionHandler.handleException(response, StatusCode.UNAUTHORIZED, MessageUtil.getMessage(StatusCode.UNAUTHORIZED));
         }
     }
 }
